Some changes that look like undoing previous changes.
	modified:   test/rmgpy/data/thermoTest.py
	modified:   test/rmgpy/kinetics/arrheniusTest.py
	modified:   test/rmgpy/statmech/rotationTest.py
	modified:   test/rmgpy/statmech/torsionTest.py

undoing some conversions of setup_method to setup_class.
Is this redone later?

+++ test/rmgpy/kinetics/arrheniusTest.py
@@ -791,10 +791,10 @@ class TestPDepArrhenius:
         """
         Tdata = np.array(
             [300, 400, 500, 600, 700, 800, 900, 1000, 1100, 1200, 1300, 1400, 1500],
-            np.float,
+            float,
         )
-        Pdata = np.array([1e4, 3e4, 1e5, 3e5, 1e6], np.float)
-        kdata = np.zeros([len(Tdata), len(Pdata)], np.float)
+        Pdata = np.array([1e4, 3e4, 1e5, 3e5, 1e6], float)
+        kdata = np.zeros([len(Tdata), len(Pdata)], float)
         for t in range(len(Tdata)):
             for p in range(len(Pdata)):
                 kdata[t, p] = self.kinetics.get_rate_coefficient(Tdata[t], Pdata[p])
